requirements.txt
google-generativeai>=0.3.0
speechrecognition>=3.10.0
pyttsx3>=2.90
gspread>=5.10.0
oauth2client>=4.1.3
pyaudio>=0.2.11
python-dotenv>=1.0.0